refactor: deduplicate PeerIdStr type (#7824)
**Motivation**

Small clean-up PR. Deduplicates three instances of `PeerIdStr` type
definition.
