+++ packages/beacon-node/src/network/core/types.ts
@@ -4,13 +4,13 @@ import {routes} from "@lodestar/api";
 import {LoggerNodeOpts} from "@lodestar/logger/node";
 import {ResponseIncoming} from "@lodestar/reqresp";
 import {phase0} from "@lodestar/types";
+import {PeerIdStr} from "../../util/peerId.js";
 import {NetworkOptions} from "../options.js";
 import {PeerAction, PeerScoreStats} from "../peers/index.js";
 import {OutgoingRequestArgs} from "../reqresp/types.js";
 import {CommitteeSubscription} from "../subnets/interface.js";
 
 export type MultiaddrStr = string;
-export type PeerIdStr = string;
 
 // Interface shared by main Network class, and all backends
 export interface INetworkCorePublic {
